+++ .github/workflows/tests.yaml
@@ -68,6 +68,7 @@ env:
 
 jobs:
   linting-coverage:
+    if: ${{ false }}
     runs-on: ubuntu-latest
     steps:
     - name: Checkout
@@ -167,6 +168,7 @@ jobs:
       tag: ${{ github.sha }}-svc-base
 
   multiple-backend:
+    if: ${{ false }}
     runs-on: ubuntu-latest
     needs: build
     env:
@@ -249,6 +251,7 @@ jobs:
       if: always()
 
   utapi-v2-tests:
+    if: ${{ false }}
     runs-on: ubuntu-latest
     needs: build
     env:
@@ -284,6 +287,7 @@ jobs:
       if: always()
 
   kmip-ft-tests:
+    if: ${{ false }}
     runs-on: ubuntu-latest
     needs: build
     env:
