+++ lib/api/bucketDeleteTagging.js
@@ -3,7 +3,7 @@ const { metadataValidateBucket } = require('../metadata/metadataUtils');
 const { pushMetric } = require('../utapi/utilities');
 const metadata = require('../metadata/wrapper');
 const util = require('node:util');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 
 /**
  * Bucket Delete Tagging - Delete a bucket's Tagging
