+++ lib/api/bucketDeleteWebsite.js
@@ -6,7 +6,7 @@ const { isBucketAuthorized } =
     require('./apiUtils/authorization/permissionChecks');
 const metadata = require('../metadata/wrapper');
 const { pushMetric } = require('../utapi/utilities');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 
 const requestType = 'bucketDeleteWebsite';
