+++ lib/api/bucketGet.js
@@ -7,7 +7,7 @@ const { metadataValidateBucket } = require('../metadata/metadataUtils');
 const collectCorsHeaders = require('../utilities/collectCorsHeaders');
 const escapeForXml = s3middleware.escapeForXml;
 const { pushMetric } = require('../utapi/utilities');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 
 const versionIdUtils = versioning.VersionID;
 const { generateToken, decryptToken }
