+++ lib/api/bucketGetTagging.js
@@ -3,7 +3,7 @@ const util = require('node:util');
 const collectCorsHeaders = require('../utilities/collectCorsHeaders');
 const { checkExpectedBucketOwner } = require('./apiUtils/authorization/bucketOwner');
 const { pushMetric } = require('../utapi/utilities');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 const { errors, s3middleware } = require('arsenal');
 const escapeForXml = s3middleware.escapeForXml;
