+++ lib/api/bucketHead.js
@@ -2,7 +2,7 @@ const collectCorsHeaders = require('../utilities/collectCorsHeaders');
 const { metadataValidateBucket } = require('../metadata/metadataUtils');
 
 const { pushMetric } = require('../utapi/utilities');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 
 /**
  * Determine if bucket exists and if user has permission to access it
