+++ lib/api/bucketPutTagging.js
@@ -7,7 +7,7 @@ const { metadataValidateBucket } = require('../metadata/metadataUtils');
 const metadata = require('../metadata/wrapper');
 const { pushMetric } = require('../utapi/utilities');
 const { checkExpectedBucketOwner } = require('./apiUtils/authorization/bucketOwner');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 const { parseTagXml } = s3middleware.tagging;
 
 /**
