+++ lib/api/initiateMultipartUpload.js
@@ -14,7 +14,7 @@ const locationConstraintCheck
     = require('./apiUtils/object/locationConstraintCheck');
 const validateWebsiteHeader = require('./apiUtils/object/websiteServing')
     .validateWebsiteHeader;
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 const { validateHeaders, compareObjectLockInformation } =
     require('./apiUtils/object/objectLockHelpers');
 const { getObjectSSEConfiguration } = require('./apiUtils/bucket/bucketEncryption');
