+++ lib/api/multiObjectDelete.js
@@ -16,7 +16,7 @@ const { isBucketAuthorized } =
 const { preprocessingVersioningDelete }
     = require('./apiUtils/object/versioning');
 const createAndStoreObject = require('./apiUtils/object/createAndStoreObject');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 const { metadataGetObject } = require('../metadata/metadataUtils');
 const { config } = require('../Config');
 const { hasGovernanceBypassHeader, checkUserGovernanceBypass, ObjectLockInfo }
