+++ lib/api/objectGetTagging.js
@@ -8,7 +8,7 @@ const { metadataValidateBucketAndObj } = require('../metadata/metadataUtils');
 const { pushMetric } = require('../utapi/utilities');
 const collectCorsHeaders = require('../utilities/collectCorsHeaders');
 const { convertToXml } = s3middleware.tagging;
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 
 /**
  * Object Get Tagging - Return tag for object
