+++ lib/api/objectHead.js
@@ -7,7 +7,7 @@ const collectCorsHeaders = require('../utilities/collectCorsHeaders');
 const collectResponseHeaders = require('../utilities/collectResponseHeaders');
 const { pushMetric } = require('../utapi/utilities');
 const { getVersionIdResHeader } = require('./apiUtils/object/versioning');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 const { getPartNumber, getPartSize, getPartCountFromMd5 } =
     require('./apiUtils/object/partInfo');
