+++ lib/api/objectPut.js
@@ -14,7 +14,7 @@ const { hasNonPrintables } = require('../utilities/stringChecks');
 const kms = require('../kms/wrapper');
 const { config } = require('../Config');
 const { setExpirationHeaders } = require('./apiUtils/object/expirationHeaders');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 const writeContinue = require('../utilities/writeContinue');
 const versionIdUtils = versioning.VersionID;
