+++ lib/api/objectPutCopyPart.js
@@ -13,7 +13,7 @@ const logger = require('../utilities/logger');
 const services = require('../services');
 const setUpCopyLocator = require('./apiUtils/object/setUpCopyLocator');
 const { metadataValidateBucketAndObj } = require('../metadata/metadataUtils');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 
 const versionIdUtils = versioning.VersionID;
 const { config } = require('../Config');
