+++ lib/api/websiteGet.js
@@ -11,7 +11,7 @@ const { isObjAuthorized, isBucketAuthorized } =
     require('./apiUtils/authorization/permissionChecks');
 const collectResponseHeaders = require('../utilities/collectResponseHeaders');
 const { pushMetric } = require('../utapi/utilities');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 
 /**
  * _errorActions - take a number of actions once have error getting obj
