+++ lib/api/websiteHead.js
@@ -9,7 +9,7 @@ const { findRoutingRule, extractRedirectInfo } =
     require('./apiUtils/object/websiteServing');
 const collectResponseHeaders = require('../utilities/collectResponseHeaders');
 const { pushMetric } = require('../utapi/utilities');
-const monitoring = require('../utilities/monitoringHandler');
+const monitoring = require('../utilities/metrics');
 const { isBucketAuthorized, isObjAuthorized } =
     require('./apiUtils/authorization/permissionChecks');
