+++ lib/server.js
@@ -5,6 +5,7 @@ const { series } = require('async');
 const arsenal = require('arsenal');
 const { RedisClient, StatsClient } = arsenal.metrics;
 const monitoringClient = require('./utilities/monitoringHandler');
+const metrics = require('./utilities/metrics');
 
 const logger = require('./utilities/logger');
 const { internalHandlers } = require('./utilities/internalHandlers');
@@ -88,7 +89,7 @@ class S3Server {
      * @returns {void}
      */
     routeRequest(req, res) {
-        monitoringClient.httpActiveRequests.inc();
+        metrics.httpActiveRequests.inc();
         const requestStartTime = process.hrtime.bigint();
 
         // disable nagle algorithm
@@ -108,11 +109,11 @@ class S3Server {
             if (req.apiMethod) {
                 labels.action = req.apiMethod;
             }
-            monitoringClient.httpRequestsTotal.labels(labels).inc();
-            monitoringClient.httpRequestDurationSeconds
+            metrics.httpRequestsTotal.labels(labels).inc();
+            metrics.httpRequestDurationSeconds
                 .labels(labels)
                 .observe(responseTimeInNs / 1e9);
-            monitoringClient.httpActiveRequests.dec();
+            metrics.httpActiveRequests.dec();
         };
         res.on('close', monitorEndOfRequest);
 
@@ -140,52 +141,6 @@ class S3Server {
             },
         };
         routes(req, res, params, logger, _config);
-    }
-
-    /**
-     * Route requests on 'admin' port
-     * @param {http.IncomingMessage} req - http request object
-     * @param {http.ServerResponse} res - http response object
-     * @returns {void}
-     */
-    routeAdminRequest(req, res) {
-        // use proxied hostname if needed
-        if (req.headers['x-target-host']) {
-            // eslint-disable-next-line no-param-reassign
-            req.headers.host = req.headers['x-target-host'];
-        }
-
-        const clientInfo = {
-            clientIP: req.socket.remoteAddress,
-            clientPort: req.socket.remotePort,
-            httpMethod: req.method,
-            httpURL: req.url,
-            endpoint: req.endpoint,
-        };
-
-        let reqUids = req.headers['x-scal-request-uids'];
-        if (reqUids !== undefined && !(reqUids.length < 128)) {
-            // simply ignore invalid id (any user can provide an
-            // invalid request ID through a crafted header)
-            reqUids = undefined;
-        }
-        const log = (reqUids !== undefined ?
-            logger.newRequestLoggerFromSerializedUids(reqUids) :
-            logger.newRequestLogger());
-        log.end().addDefaultFields(clientInfo);
-
-        log.debug('received admin request', clientInfo);
-
-        switch (req.url) {
-        case '/live':
-        case '/ready':
-            healthcheckHandler(clientInfo.clientIP, req, res, log, statsClient);
-            break;
-
-        default:
-            monitoringClient.monitoringHandler(clientInfo.clientIP, req, res, log);
-            break;
-        }
     }
 
     /**
@@ -299,18 +254,67 @@ class S3Server {
                 this._startServer(this.routeRequest, _config.port);
             }
 
-            // Start metrics server(s)
-            if (_config.metricsListenOn.length > 0) {
-                _config.metricsListenOn.forEach(item => {
-                    this._startServer(this.routeAdminRequest, item.port, item.ip);
-                });
-            } else {
-                this._startServer(this.routeAdminRequest, _config.metricsPort);
-            }
-
             this.started = true;
         });
     }
+}
+
+/**
+ * Route requests on 'admin' port
+ * @param {http.IncomingMessage} req - http request object
+ * @param {http.ServerResponse} res - http response object
+ * @returns {void}
+ */
+function _routeAdminRequest(req, res) {
+    // use proxied hostname if needed
+    if (req.headers['x-target-host']) {
+        // eslint-disable-next-line no-param-reassign
+        req.headers.host = req.headers['x-target-host'];
+    }
+
+    const clientInfo = {
+        clientIP: req.socket.remoteAddress,
+        clientPort: req.socket.remotePort,
+        httpMethod: req.method,
+        httpURL: req.url,
+        endpoint: req.endpoint,
+    };
+
+    let reqUids = req.headers['x-scal-request-uids'];
+    if (reqUids !== undefined && !(reqUids.length < 128)) {
+        // simply ignore invalid id (any user can provide an
+        // invalid request ID through a crafted header)
+        reqUids = undefined;
+    }
+    const log = (reqUids !== undefined ?
+        logger.newRequestLoggerFromSerializedUids(reqUids) :
+        logger.newRequestLogger());
+    log.end().addDefaultFields(clientInfo);
+
+    log.debug('received admin request', clientInfo);
+
+    switch (req.url) {
+    case '/live':
+    case '/ready':
+        healthcheckHandler(clientInfo.clientIP, req, res, log, statsClient);
+        break;
+
+    default:
+        monitoringClient.monitoringHandler(clientInfo.clientIP, req, res, log);
+        break;
+    }
+}
+
+function setupMetricsServer() {
+    const WebServer = arsenal.network.http.server;
+    const metricsServer = new WebServer(_config.metricsPort, logger);
+    metricsServer.onRequest(_routeAdminRequest);
+    if (_config.https) {
+        const { ca, cert, key } = _config.https;
+        metricsServer.setHttps(cert, key, ca);
+    }
+    // start
+    metricsServer.start();
 }
 
 function main() {
@@ -362,11 +366,16 @@ function main() {
                 workerPid: worker.process.pid,
             });
         });
+        // setup metrics server in master, prom-client will not aggregate metrics
+        // from workers by default, use prom-client's aggregate registry to collect
+        // from all workers!
+        setupMetricsServer();
     } else {
         const server = new S3Server(cluster.worker);
         server.initiateStartup(logger.newRequestLogger());
+        // collect default metrics from the workers (will be aggregated in master)
+        monitoringClient.collectDefaultMetrics();
     }
-    monitoringClient.collectDefaultMetrics({ timeout: 5000 });
 }
 
 module.exports = main;
